chore: add TypeDoc, rename fill to flood
