+++ package.json
@@ -23,7 +23,6 @@
   "files": [
     "/dist"
   ],
-  "dependencies": {},
   "devDependencies": {
     "@babel/core": "7.17.2",
     "@babel/preset-env": "7.16.11",
@@ -38,6 +37,7 @@
     "ts-jest": "27.1.3",
     "ts-loader": "8.3.0",
     "ts-node": "10.5.0",
+    "typedoc": "0.22.11",
     "typescript": "4.5.5",
     "webpack": "5.68.0",
     "webpack-cli": "4.9.2"
@@ -49,7 +49,8 @@
     "test-cov": "jest --coverage",
     "test-build": "npm test && npm build",
     "fmt": "prettier --config .prettierrc 'src/**/*.ts' --write",
-    "fmt-check": "prettier --config .prettierrc 'src/**/*.ts' --check"
+    "fmt-check": "prettier --config .prettierrc 'src/**/*.ts' --check",
+    "docs": "typedoc"
   },
   "husky": {
     "hooks": {
@@ -59,5 +60,7 @@
   "lint-staged": {
     "src/**/*.ts": "prettier --write --config .prettierrc"
   },
-  "engines": {"node": "16"}
+  "engines": {
+    "node": "16"
+  }
 }
