--- src/helpers/types.ts
@@ -1,4 +1,0 @@
-export type Position = {
-    x: number;
-    y: number;
-};
