+++ src/helpers/types/index.ts
@@ -1,0 +1,7 @@
+/** Represents coordinates of any two dimensional array. */
+type Position = {
+    x: number;
+    y: number;
+};
+
+export { Position };
