+++ src/index.ts
@@ -1,8 +1,9 @@
 import * as FOV from './fov/index';
 import * as Line from './line/index';
-import * as Fill from './fill/index';
+import * as Flood from './flood/index';
 import * as Path from './path/index';
 import * as PositionHelpers from './helpers/position';
+import * as Types from './helpers/types';
 
-export { FOV, Line, Fill, Path, PositionHelpers };
+export { FOV, Line, Flood, Path, PositionHelpers, Types };
 export { default as Draw } from './draw/draw';
